+++ libs/vre/shared/app-resource-properties/src/index.ts
@@ -26,7 +26,7 @@ export * from './lib/existing-property-value.component';
 export * from './lib/property-value-action-bubble.component';
 export * from './lib/property-value.component';
 export * from './lib/form-value-array.type';
-export * from './lib/resource-properties.components';
+export * from './resource-properties.components';
 export * from './lib/resource-payloads-mapping';
 export * from './lib/create-resource-dialog.component';
 export * from './lib/create-resource-form.component';
@@ -47,3 +47,4 @@ export * from './lib/incoming-link.interface';
 export * from './lib/incoming.service';
 export * from './lib/resource-info-bar.component';
 export * from './lib/sortByKeys';
+export * from './lib/properties-display.component';
