+++ libs/vre/shared/app-resource-properties/src/lib/properties-display-incoming-link.service.ts
